Move differenct functionality to proper files
