+++ main.js
@@ -1,5 +1,3 @@
-const passport = require('passport');
-const GitHubStrategy = require('passport-github').Strategy;
 const express = require('express');
 const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
@@ -9,9 +7,14 @@ const _ = require('lodash');
 const graphql = require('@octokit/graphql');
 const util = require('util');
 
+const passport = require('./auth/passport');
+
 const pullRequestQueryPart = require('./queries-part/pull-request');
 const repositoriesContributedToQueryPart = require('./queries-part/repositories-contributed-to');
 const repositoriesQueryPart = require('./queries-part/repositories');
+
+const repositoriesContributedToMapper = require('./mapper/repositories-contributed-to');
+const repositoriesMapper = require('./mapper/repositories');
 
 const app = express();
 app.use(morgan('dev'));
@@ -22,35 +25,9 @@ app.use(expressSession({
     resave: true,
     saveUninitialized: true
 }));
+
 app.use(passport.initialize());
 app.use(passport.session());
-const APP_PORT = process.env.PORT || 4100;
-const APP_URL = process.env.APP_URL || 'http://127.0.0.1';
-const GITHUB_CLIENT_CALLBACK = process.env.APP_URL ? `${APP_URL}auth/github/callback` : `${APP_URL}:${APP_PORT}/auth/github/callback`;
-const GITHUB_CLIENT_ID = process.env.GITHUB_CLIENT_ID || '46cd8365e525e9c25d44';
-const GITHUB_CLIENT_SECRET = process.env.GITHUB_CLIENT_SECRET || '9ae7aad707674bf89e08f201529b3ec63b6f0513';
-
-passport.use(new GitHubStrategy({
-        clientID: GITHUB_CLIENT_ID,
-        clientSecret: GITHUB_CLIENT_SECRET,
-        callbackURL: GITHUB_CLIENT_CALLBACK
-    },
-    (accessToken, refreshToken, profile, cb) => {
-        const user = {
-            token: null,
-            githubProfile: null
-        };
-        user.token = accessToken;
-        user.githubProfile = _.omit(profile, ['_raw', '_json']);
-        return cb(null, user);
-    }
-));
-passport.serializeUser(function (user, done) {
-    done(null, user);
-});
-passport.deserializeUser(function (user, done) {
-    done(null, user);
-});
 
 app.get('/', (req, res) => {
     if (req.user) {
@@ -67,18 +44,8 @@ app.get('/', (req, res) => {
         }).then((data) => {
             console.log(util.inspect(data, { showHidden: true, depth: null }));
             req.user.pullRequests = _.map(data.viewer.pullRequests.edges, 'node');
-            req.user.repositoriesContributedTo = _.map(data.viewer.repositoriesContributedTo.edges, (repositoryContributedTo) => {
-                return {
-                    name: repositoryContributedTo.node.name,
-                    commits: _.map(repositoryContributedTo.node.ref.target.history.edges, 'node.messageHeadline')
-                };
-            });
-            req.user.repositories = _.map(data.viewer.repositories.edges, (repository) => {
-                return {
-                    name: repository.node.name,
-                    commits: _.map(repository.node.ref.target.history.edges, 'node.messageHeadline')
-                };
-            });
+            req.user.repositoriesContributedTo = _.map(data.viewer.repositoriesContributedTo.edges, repositoriesContributedToMapper);
+            req.user.repositories = _.map(data.viewer.repositories.edges, repositoriesMapper);
             res.json(req.user);
         });
     } else {
@@ -89,4 +56,4 @@ app.get('/', (req, res) => {
 app.get('/loginFailure', (req, res) => res.status(403).json({status: 403}));
 app.get('/auth/github', passport.authenticate('github'));
 app.get('/auth/github/callback', passport.authenticate('github', {failureRedirect: '/loginFailure'}), (req, res) => res.redirect('/'));
-app.listen(APP_PORT, '0.0.0.0', () => console.log(`Example app listening on port ${APP_PORT}!`));
+app.listen(process.env.PORT || 4100, '0.0.0.0', () => console.log(`Application started!`));
